Enable GitLab CI Pipelines
This includes not only the mandatory .gitlab-ci.yml file containing
the pipeline configuration but also an Emacs Lisp file for exporting
the blog post and pages from the content-org files to markdown using
Emacs and ox-hugo.

It was also necessary to remove thr orgit links from the "My Emacs
Package of the week: orgit" post since these required that the linked
repository is at the given path which is not possible in the CI build.
Well, it is /possible/ but there is no good way to do it.

Additional the .gitmodules file was updated to use a relative local
path instead of and SSH URL since GitLab can obviously not clone form
an SSH URL. This may make things more difficult for local setups but I
think I can live with the one additional clone command.