🔥 Remove "Support Ukraine" banner
Not  because it  is  "old".  But because,  sadly,  there  are so  many
cruelties going  on at  the moment  that it no  longer feels  right to
highlight just one. And making every page  of this site a long list of
current disasters is also not a good way forward.

Let's all hope for a brighter tomorrow!